[FACU] *)corregido en observables.c el valor de retorno de las energias a double. *) implementada la funcion para grabar las posiciones de las particulas en formato pdb.

+++ src/observables.c
@@ -5,7 +5,7 @@
 #include <time.h>
 #include <math.h>
 
-float Energia_Potencial(double* pos_vel, int N, double* LUT_P, int Ntabla, double L){
+double Energia_Potencial(double* pos_vel, int N, double* LUT_P, int Ntabla, double L){
   double rij = 0;
   double V = 0;
   for(int i=1;i<N;i++){
@@ -17,7 +17,7 @@ float Energia_Potencial(double* pos_vel, int N, double* LUT_P, int Ntabla, doubl
   return V;
 }
 
-float Energia_Cinetica(double* pos_vel, int N, double m){
+double Energia_Cinetica(double* pos_vel, int N, double m){
   double Ecin = 0;
   for(int i=0;i<N;i++){
     Ecin = Ecin + 0.5*m*(pos_vel[i+3*N]*pos_vel[i+3*N]+pos_vel[i+4*N]*pos_vel[i+4*N]+pos_vel[i+5*N]*pos_vel[i+5*N]);
